rozdzeilenie odpowiedzialnosci
